model_deployment/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# ML dependencies
torch>=2.2.0
transformers>=4.44.0
peft>=0.10.0
accelerate>=0.27.0
safetensors>=0.4.1
numpy<2.0.0
google-generativeai>=0.3.0
Pillow>=9.0.0
requests>=2.31.0
